refactor: enables HTTPS for local dev purposes
