+++ src/bootstrap/index.ts
@@ -1,3 +1,5 @@
+import { existsSync, readFileSync } from 'fs'
+
 import { ValidationPipe } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { NestFactory } from '@nestjs/core'
@@ -9,9 +11,10 @@ import buildLogger from 'lib/logger'
 import BootstrapModule from './module'
 
 async function bootstrap() {
+  const https = buildHttpsConfig()
   const app = await NestFactory.create<NestFastifyApplication>(
     BootstrapModule,
-    new FastifyAdapter(),
+    new FastifyAdapter({ https }),
   )
 
   const configService: ConfigService = app.get(ConfigService)
@@ -31,6 +34,20 @@ async function bootstrap() {
 
   await app.listen(configService.get('port'), '0.0.0.0')
   logger.log(`Started server listening to port ${appConfig.port.toString()}`)
+}
+
+const buildHttpsConfig = () => {
+  const isHttps =
+    existsSync(appConfig.https.credentials.key) && existsSync(appConfig.https.credentials.cert)
+
+  const httpsConfig = isHttps
+    ? {
+        key: readFileSync(appConfig.https.credentials.key),
+        cert: readFileSync(appConfig.https.credentials.cert),
+      }
+    : {}
+
+  return httpsConfig
 }
 
 export default bootstrap
